refactor: Decouple load balance metrics collection from DividePlugin
