+++ shenyu-common/src/main/java/org/apache/shenyu/common/constant/Constants.java
@@ -1069,6 +1069,11 @@ public interface Constants {
     String GB = "GB";
 
     /**
+     * The request begin time.
+     */
+    String REQUEST_BEGIN_TIME = "requestBeginTime";
+
+    /**
      * String q.
      */
     default void findConstants() {
